backend/requirements.txt
# /root/nubla-siem/backend/requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
psycopg2-binary==2.9.7
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.4.0
python-dotenv==1.0.1
elasticsearch==8.9.0
redis==5.0.1
confluent-kafka==2.5.3
python-multipart==0.0.18
aiohttp==3.10.11
slowapi==0.1.9
pyotp==2.9.0
qrcode==8.0
pillow==10.4.0
transformers==4.48.0
scikit-learn==1.5.0
tensorflow-cpu==2.14.0
torch==2.6.0
esper==2.5
setuptools==70.0.0
starlette==0.40.0
numpy==1.26.4
pydantic==2.8.2
pydantic_settings==2.5.2